refactor(core): consolidate duplicated render/provider helpers into shared modules
Five clusters of copy-pasted code, grouped by real consumer rather than
forced into one grab-bag:

- badges/satori-fonts.ts (new): findFontsDir + FONT_CONFIG + getFonts, was
  byte-identical in render.tsx and render-group.tsx. Kept separate from
  fonts.ts (which shapes the same TTFs as Uint8Array[] for resvg PNG).
- render.tsx now exports luminance/rgba; render-group.tsx imports them.
- badges/svg-text-utils.ts (new): esc/r2/clamp, triplicated across the
  chart/header/sponsors hand-built-SVG renderers.
- provider-fetch.ts now exports isRateLimitResponse; github.ts and
  starhistory.ts import it instead of each defining the same GitHub
  403-quota detection.
- providers/coverage-color.ts (new): coveragePctAndColor, the shared
  coverage-%-to-color threshold ladder used by codecov and coveralls.

render-chart's own rgba is intentionally left in place — it emits a
spaced rgba(r, g, b, a) for inline SVG, distinct from render.tsx's
compact form for Satori; merging would change output bytes.

Verified: tsc clean, full test suite green (288/10 core), next build,
and a live screenshot pass confirming badge/group/chart/header render
identically.
